requirements.txt
tensorflow
gradio
pillow
opencv-python
numpy
matplotlib
scikit-learn